Rename /components/user to /components/profile, move login-form.tsx to a new /components/auth folder
