+++ client/src/app/(user)/profile/banking-information/page.tsx
@@ -1,5 +1,5 @@
 import { getUser } from '@/api';
-import BankingInformationForm from '@/components/user/BankingInformationForm';
+import BankingInformationForm from '@/components/profile/BankingInformationForm';
 
 async function BankingInformationPage() {
   const user = await getUser(1);
